Add data sending
uses Glitch server right now because I don't have the money to buy a server from a provider. If anyone has a better idea or wants to donate to have an actually good server for the data transmission feel free to hit me up. If I suddenly become rich I'll set up a server for this.