Added TikTok support
Added TikTok support. It only works partially both because of the TikTok UI being stupid and not allowing extension context menu to show (or at least it's complicated enough that I don't care to try). There's also a weird bug that only colours the time since upload on TikTok videos and not the user.